BREAKING CHANGE: Remove object destructuring
