+++ .gitignore
@@ -2,3 +2,4 @@ package-lock.json
 node_modules
 coverage
 dist
+.DS_Store
