+++ CHANGELOG.md
@@ -1,63 +1,56 @@
 ## [2.0.0](https://github.com/coly010/notry/compare/v1.1.3...v2.0.0) (YYY-MM-DD)
 
-
-### Breaking Changes (no-try/tuple)
+### Breaking Changes
 
 The Node.js `callback(err, res)` pattern makes sure that the error object is explicitly available and needs to be checked first before expecting the response object. Hence, ignoring the error object requires intentionally ignoring the first argument.
 
-With the goal of making it simpler to adopt better error handling practices, the de-facto approach of this library is now `[err, res]` compared to the `[res, err]` format in versions `v1.1.x`.
+With the goal of making it simpler to adopt better error handling practices, the de-facto approach of this library is now `[err, res]` compared to the `[res, err]` and `{ error, result }` formats in previous versions.
 
-* invert tuple ordering ([to be updated](https://github.com/coly010/notry/commit/d4fbd09238b49b2c80c396bee8ec3e1b27ab43f8))
+- invert tuple ordering ([to be updated](https://github.com/coly010/notry/commit/d4fbd09238b49b2c80c396bee8ec3e1b27ab43f8))
+- remove object destructuring ([to be updated](https://github.com/coly010/notry/commit/d4fbd09238b49b2c80c396bee8ec3e1b27ab43f8))
 
 ## [1.1.3](https://github.com/coly010/notry/compare/v1.1.2...v1.1.3) (2020-01-27)
 
-
 ### Bug Fixes
 
-* export tuple correctly ([d4fbd09](https://github.com/coly010/notry/commit/d4fbd09238b49b2c80c396bee8ec3e1b27ab43f8))
-* export tuple correctly ([5377b9b](https://github.com/coly010/notry/commit/5377b9b721922892ff590712097bee92efae7a2b))
+- export tuple correctly ([d4fbd09](https://github.com/coly010/notry/commit/d4fbd09238b49b2c80c396bee8ec3e1b27ab43f8))
+- export tuple correctly ([5377b9b](https://github.com/coly010/notry/commit/5377b9b721922892ff590712097bee92efae7a2b))
 
 ## [1.1.2](https://github.com/coly010/notry/compare/v1.1.1...v1.1.2) (2020-01-27)
 
-
 ### Bug Fixes
 
-* tuple export path ([d4e9b4c](https://github.com/coly010/notry/commit/d4e9b4c26605243b5d9de14a5d4c25ce4b713905))
-* tuple export path ([fce9468](https://github.com/coly010/notry/commit/fce9468d0be42c7027440c239d92f5d8a269414a))
+- tuple export path ([d4e9b4c](https://github.com/coly010/notry/commit/d4e9b4c26605243b5d9de14a5d4c25ce4b713905))
+- tuple export path ([fce9468](https://github.com/coly010/notry/commit/fce9468d0be42c7027440c239d92f5d8a269414a))
 
 ## [1.1.1](https://github.com/coly010/notry/compare/v1.1.0...v1.1.1) (2020-01-27)
 
-
 ### Bug Fixes
 
-* tuple folder path for build ([17fd35d](https://github.com/coly010/notry/commit/17fd35d79f12cbfee16c51b44c9fa2c018c6fe00))
-* tuple folder path for build ([11bf028](https://github.com/coly010/notry/commit/11bf028e25dca779291266a7a05e4ec47c4c98e8))
+- tuple folder path for build ([17fd35d](https://github.com/coly010/notry/commit/17fd35d79f12cbfee16c51b44c9fa2c018c6fe00))
+- tuple folder path for build ([11bf028](https://github.com/coly010/notry/commit/11bf028e25dca779291266a7a05e4ec47c4c98e8))
 
 # [1.1.0](https://github.com/coly010/notry/compare/v1.0.1...v1.1.0) (2020-01-27)
 
-
 ### Features
 
-* add support for tuple return [#3](https://github.com/coly010/notry/issues/3) ([e86c65e](https://github.com/coly010/notry/commit/e86c65e8c7e18d053ea07666f22a72ac50ee6efa))
-* add support for tuple return [#3](https://github.com/coly010/notry/issues/3) ([29052fa](https://github.com/coly010/notry/commit/29052fa7c6419d6d8fcefd9a5a41108336687b7b))
+- add support for tuple return [#3](https://github.com/coly010/notry/issues/3) ([e86c65e](https://github.com/coly010/notry/commit/e86c65e8c7e18d053ea07666f22a72ac50ee6efa))
+- add support for tuple return [#3](https://github.com/coly010/notry/issues/3) ([29052fa](https://github.com/coly010/notry/commit/29052fa7c6419d6d8fcefd9a5a41108336687b7b))
 
 ## [1.0.1](https://github.com/coly010/notry/compare/v1.0.0...v1.0.1) (2019-11-30)
 
-
 ### Bug Fixes
 
-* ci build failing ([c84a082](https://github.com/coly010/notry/commit/c84a0825b8ce94fae3223f9bf86a6fb04c859e0b))
+- ci build failing ([c84a082](https://github.com/coly010/notry/commit/c84a0825b8ce94fae3223f9bf86a6fb04c859e0b))
 
 # 1.0.0 (2019-11-30)
 
-
 ### Bug Fixes
 
-* build and publish script ([4a75d11](https://github.com/coly010/notry/commit/4a75d111273f5902d45c8b8d04a078d398fe6bc5))
-* semantic release deps ([670e602](https://github.com/coly010/notry/commit/670e602aa8c5c6083a189d245bd46aba0f1a7fdd))
-
+- build and publish script ([4a75d11](https://github.com/coly010/notry/commit/4a75d111273f5902d45c8b8d04a078d398fe6bc5))
+- semantic release deps ([670e602](https://github.com/coly010/notry/commit/670e602aa8c5c6083a189d245bd46aba0f1a7fdd))
 
 ### Features
 
-* add no try source ([5f8b3f6](https://github.com/coly010/notry/commit/5f8b3f6390e9dd852a5f097eadc340282098e4aa))
-* **notry:** add source functions and tests ([3a3e560](https://github.com/coly010/notry/commit/3a3e560e3927df139a5dc5a81e7525c5ed61d1b8))
+- add no try source ([5f8b3f6](https://github.com/coly010/notry/commit/5f8b3f6390e9dd852a5f097eadc340282098e4aa))
+- **notry:** add source functions and tests ([3a3e560](https://github.com/coly010/notry/commit/3a3e560e3927df139a5dc5a81e7525c5ed61d1b8))
