+++ README.md
@@ -7,7 +7,7 @@
   <a href="https://www.npmjs.com/package/no-try"><img src="https://img.shields.io/badge/npm-no--try-brightgreen.svg" /></a>
   <a href="https://www.npmjs.com/package/no-try"><img src="https://img.shields.io/npm/v/no-try.svg" /></a>
   <a href="https://www.npmjs.com/package/no-try"><img src="https://img.shields.io/npm/dt/no-try.svg" /></a>
-  <a href="https://www.npmjs.com/package/no-try"><img src="https://img.shields.io/travis/coly010/notry.svg" /></a>
+  <a href="https://www.npmjs.com/package/no-try"><img src="https://img.shields.io/travis/coly010/no-try.svg" /></a>
   <a href="https://www.npmjs.com/package/no-try"><img src="https://img.shields.io/npm/l/no-try.svg" /></a>
 </p>
 
@@ -17,100 +17,17 @@ Working in a code base where you can expect methods to `throw` can lead to situa
 
 `no-try` tackles this by removing the `try-catch` to an external method, whilst allowing the flexibility of handling the error thrown appropriately and getting access to the return value of the method that will throw. 🤘🤘
 
-`no-try` now supports both Object Destructuring, as well as Array Destructuring, allowing for more flexible naming of your variables.
-
 ## 🔧 Installation
 
 `npm install --save no-try`
 
 ## 🎸 Usage
 
-To get up and running quickly, simply install the package:
-
-`npm install no-try`
-
-Then use it in your code:
-
-Object Destructuring:
-
-JS:
+#### JavaScript
 
 ```js
 const noTry = require("no-try").noTry;
 const noTryAsync = require("no-try").noTryAsync;
-
-// Without a custom error handler
-const { result, error } = noTry(() => myThrowableMethod());
-
-// With a custom error handler
-const { result, error } = noTry(
-  () => myThrowableMethod(),
-  error => {
-    console.log(error);
-  }
-);
-
-// Handle methods that return a Promise without a custom error handler
-const { result, error } = await noTryAsync(() => myAsyncThrowableMethod());
-
-// Handle methods that return a Promise with a custom error handler
-const { result, error } = await noTryAsync(
-  () => myAsyncThrowableMethod(),
-  error => {
-    console.log(error);
-  }
-);
-
-// Use result
-if (error) {
-  // Show error alert
-}
-
-sendMyResultToMethod(result);
-```
-
-TS:
-
-```ts
-import { noTry, noTryAsync } from "no-try";
-
-// Without a custom error handler
-const { result, error } = noTry(() => myThrowableMethod());
-
-// With a custom error handler
-const { result, error } = noTry(
-  () => myThrowableMethod(),
-  error => {
-    console.log(error);
-  }
-);
-
-// Handle methods that return a Promise without a custom error handler
-const { result, error } = await noTryAsync(() => myAsyncThrowableMethod());
-
-// Handle methods that return a Promise with a custom error handler
-const { result, error } = await noTryAsync(
-  () => myAsyncThrowableMethod(),
-  error => {
-    console.log(error);
-  }
-);
-
-// Use result
-if (error) {
-  // Show error alert
-}
-
-sendMyResultToMethod(result);
-```
-
-Array Destructuring (Note: the `/tuple` in the import path is required):
-
-JS:
-
-```js
-const noTry = require("no-try/tuple").noTry;
-const noTryAsync = require("no-try/tuple").noTryAsync;
 
 // Without a custom error handler
 const [error, result] = noTry(() => myThrowableMethod());
@@ -124,10 +41,10 @@ const [err, res] = noTry(
 );
 
 // Handle methods that return a Promise without a custom error handler
-const [err2, item] = await noTryAsync(() => myAsyncThrowableMethod());
+const [err2, res2] = await noTryAsync(() => myAsyncThrowableMethod());
 
 // Handle methods that return a Promise with a custom error handler
-const [err3, res2] = await noTryAsync(
+const [err3, res3] = await noTryAsync(
   () => myAsyncThrowableMethod(),
   error => {
     console.log(error);
@@ -142,10 +59,10 @@ if (error || err || err2 || err3) {
 sendMyResultToMethod(result);
 ```
 
-TS:
+#### TypeScript
 
 ```ts
-import { noTry, noTryAsync } from "no-try/tuple";
+import { noTry, noTryAsync } from "no-try";
 
 // Without a custom error handler
 const [error, result] = noTry(() => myThrowableMethod());
@@ -159,10 +76,10 @@ const [err, res] = noTry(
 );
 
 // Handle methods that return a Promise without a custom error handler
-const [err2, item] = await noTryAsync(() => myAsyncThrowableMethod());
+const [err2, res2] = await noTryAsync(() => myAsyncThrowableMethod());
 
 // Handle methods that return a Promise with a custom error handler
-const [err3, res2] = await noTryAsync(
+const [err3, res3] = await noTryAsync(
   () => myAsyncThrowableMethod(),
   error => {
     console.log(error);
