+++ src/index.ts
@@ -1,11 +1,11 @@
-export type NoTryResult<T> = { result: T; error: Error };
+export type NoTryTupleResult<T> = [Error, T];
 
-export function noTry<T>(fn: () => T, handleErr: (error: Error) => void = error => null): NoTryResult<T> {
-  const result: NoTryResult<T> = { result: null, error: null };
+export function noTry<T>(fn: () => T, handleErr: (error: Error) => void = error => null): NoTryTupleResult<T> {
+  const result: NoTryTupleResult<T> = [null, null];
   try {
-    result.result = fn();
+    result[1] = fn();
   } catch (err) {
-    result.error = err;
+    result[0] = err;
     handleErr(err);
   }
   return result;
@@ -14,12 +14,12 @@ export function noTry<T>(fn: () => T, handleErr: (error: Error) => void = error 
 export async function noTryAsync<T>(
   fn: () => Promise<T>,
   handleErr: (error: Error) => void = error => null
-): Promise<NoTryResult<T>> {
-  const result: NoTryResult<T> = { result: null, error: null };
+): Promise<NoTryTupleResult<T>> {
+  const result: NoTryTupleResult<T> = [null, null];
   try {
-    result.result = await fn();
+    result[1] = await fn();
   } catch (err) {
-    result.error = err;
+    result[0] = err;
     handleErr(err);
   }
   return result;
